Change maximum difference in team sizes from 1 to 0.
Where faction populations are not equal, there are from the larger
faction always or almost always more players queuing, which means
that the larger faction team is always or almost always as much
larger than the small faction team as permitted by the maximum
difference in team sizes.

A difference of only one player is enough for the larger team to
always or almost always win.

On LH, Horde is the larger faction, on Darrow, Alliance, and it
is seen that on LH, the Horde team is always or almost always
one larger than the Alliance team, and vis-verse on Darrow.

This has a feedback effect, where the smaller faction always or
almost always loses, and so its players stop queuing.

In short, even a difference of one means that BGs always or
almost always lead one faction to win, consistently, on a
per-server basis.

The existing server enforced team size difference of one was
introduced presumably to ensure the teams are not so unbalanced
that one team will inevitably win; but what has been done, with
a limit of one, does not in fact achieve the intended purpose.  A
maximum difference of zero should do so.

Please see issue 1071 for discussion.

https://github.com/LightsHope/server/issues/1071